Provide helpers to be used in view specs
This brings versions of the controller spec helpers `sign_in` and
`sign_in_as` to view specs and does so in a way that avoids people
getting [errors from verified partial doubles][1]. The downside of this
approach is that it changes the subject under test (the view) by
introducing these helper methods directly on it. I don't think that's
much different than controller helper methods, though.

There is potential for confusion among users that might expect calling
`sign_in` or `sign_in_as` would come with all of the same side effects
that calling these methods from a controller or controller spec would.
In reality, they are just fancy ways to set the `current_user` local
variable and have no side effects. I feel that the potential confusion
from this has less impact than having to remember different helper
method names depending on whether you are in a controller spec or a view
spec.

As part of this change, I moved the existing
`Clearance::Testing::Helpers` module to
`Clearance::Testing::ControllerHelpers` and deprecated the old name.
This is a more accurate name and allows us to isolate the `ViewHelpers`
which are not useful in test_unit.

[1]: https://github.com/rspec/rspec-rails/issues/1076